Bump version to 3.0.0
Been a while since we published 2.0.0. The reason we are bumping this
gem straight to 3.0.0 is due to the Ruby versions that have been
dropped but also included. More information will be found in a
changelog update coming up.

https://github.com/twingly/twingly-url_cache/compare/v2.0.0...master

On another note:

twingly-url_cache prior to this commit forces the dependency dalli[^1]
to stay between >= 2.7.0 and < 2.8.0.

Recent CVE[^2] in dalli has affected versions until < 3.2.3.

Therefore it will be good to roll out this change:
https://github.com/twingly/twingly-url_cache/pull/21/files

[^1]: https://github.com/petergoldstein/dalli
[^2]: https://github.com/advisories/GHSA-3xg8-cc8f-9wv2